backend/requirements.txt
# Core Django
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==24.2
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Async / Queue
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.6.0
gevent==24.2.1
redis==5.0.3

# HTTP
requests==2.31.0
urllib3==2.2.1

# M-Pesa / Auth
cryptography==42.0.5

# Notifications
africastalking==1.2.8

# Production
gunicorn==22.0.0

# Utils
Pillow==10.3.0
python-dateutil==2.9.0
